% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalconflict-data.R
\docType{data}
\name{caprat}
\alias{caprat}
\title{Composite Index of National Capability (CINC) Ratio Network}
\format{
A matrix with 189 columns and 189 rows.
}
\source{
Gleditsch's website: \url{http://ksgleditsch.com/publ.html}
}
\usage{
data(caprat)
}
\description{
An undirected network where ties are the logged ratio of Composite Index of National Capability (CINC) scores, with the stronger country's CINC score divided by the weaker country's. This network data is adapted from Gleditsch, Salehyan, and Schultz (2008). The network includes 189 states in the international system during the post-Cold War period (1989-1995) except Samoa. The row and column names are country codes from the Correlates of War project.
}
\references{
Replication materials for Gleditsch, KS, I Salehyan, and K Schultz, 2008, "Fighting at home, fighting abroad: How civil wars lead to international disputes." \emph{Journal of Conflict Resolution} 52(4): 479-506. \url{https://doi.org/10.1177/0022002707313305}
}
\keyword{datasets}
