% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer_utilities.R
\name{check.multilayer}
\alias{check.multilayer}
\title{Check Multilayer}
\usage{
check.multilayer(nw)
}
\arguments{
\item{nw}{a \code{network} object.}
}
\value{
Silently returns \code{TRUE} or \code{FALSE} indicating whether layer membership has been properly supplied to the network object; 
also prints a confirmation in the console.
}
\description{
Check whether layer membership has been properly supplied to the network object.
}
\details{
In order for multilayer terms to be used in the ERGM, the network object requires a vertex attribute named \code{'layer.mem'} to 
identify using numeric values which layer each node belongs to.
\code{check.multilayer} checks whether the supplied network has this vertex attribute and whether it is in the correct format.
}
