% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemg-data.R
\docType{data}
\name{infrep}
\alias{infrep}
\title{ChemG Influence Network}
\format{A matrix with 30 columns and 30 rows.}
\source{
Dataverse: \url{http://hdl.handle.net/1902.1/17004}
}
\usage{
data(infrep)
}
\description{
A directed network of all 30 ChemG actors where the binary directed ties indicate the row actor cited the column actor as influential.
}
\references{
Replication materials for Leifeld, P and V Schneider, 2012, Information Exchange in Policy Networks. \emph{American Journal of Political Science} 53(3): 731-744. \url{http://dx.doi.org/10.1111/j.1540-5907.2011.00580.x}
}
\keyword{datasets}
