% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalconflict-data.R
\docType{data}
\name{intoverlap}
\alias{intoverlap}
\title{Network of Overlapping Membership in the International System (1989-1995)}
\format{A matrix with 189 columns and 189 rows.}
\source{
Correlates of War website: \url{http://www.correlatesofwar.org/data-sets/state-system-membership}
}
\usage{
data(intoverlap)
}
\description{
An undirected network where ties indicate the two countries' existence overlapped during the 1989-1995 period. This network data is adapted from the Correlates of War State System data set (v2016). The network includes 189 states in the international system during the post-Cold War period (1989-1995) except Samoa. The row and column names are country codes from the Correlates of War project.
}
\references{
\emph{Correlates of War Project}, 2017, "State System Membership List, v2016."
}
\keyword{datasets}
