% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalconflict-data.R
\docType{data}
\name{intraconf}
\alias{intraconf}
\title{Intrastate Conflict Network}
\format{A matrix with 544 columns and 189 rows.}
\source{
UCDP website: \url{https://ucdp.uu.se/downloads/}
}
\usage{
data(intraconf)
}
\description{
An undirected network of conflict between one state and one nonstate actor. This network data is adapted from the Uppsala Conflict Data Program (UCDP) Organized Conflict data set (2017, v17.1). The network includes 189 states in the international system during the post-Cold War period (1989-1995) except Samoa, and 544 nonstate actors that were involved in any kind of organized conflict in the 1984-2000 period. The row names are Correlates of War country codes and column names are actor codes from UCDP.
}
\references{
Allansson, M, E Melander, and L Themner. 2017, "Organized violence, 1989-2016." \emph{Journal of Peace Research} 54(4): 727-742. \url{https://doi.org/10.1177/0022343317718773}
}
\keyword{datasets}
