% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer.ergm.R
\name{multilayer.ergm}
\alias{multilayer.ergm}
\title{multilayer.ergm: exponential random graph models for multilayer networks}
\description{
This package extends exponential random graph models to applications with multilayer networks.
}
