% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalconflict-data.R
\docType{data}
\name{nodeatts}
\alias{nodeatts}
\title{Global Conflict Actor Attributes}
\format{A data frame with 733 rows and 7 columns:
\describe{
\item{name}{The actor's name.}
\item{interstate}{An indicator for whether the actor is engaged in interstate conflict. Nonstate actors by definition cannot be engaged in interstate conflict.}
\item{intrastate}{An indicator for whether the actor is engaged in intrastate conflict.}
\item{nonstate}{An indicator for whether the actor is engaged in nonstate conflict. States by definition cannot be engaged in nonstate conflict.}
\item{dem}{An indicator for whether the actor is a democracy. All nonstate actors are coded as nondemocracies.}
\item{cinc}{The actor's Composite Index of National Capability (CINC) score. All nonstate actors are coded to have a CINC score of zero.}
\item{trans}{An indicator for whether the actor is a transitional regime. All nonstate actors are coded to not be transitional regimes.}
}}
\source{
Gleditsch's website: \url{http://ksgleditsch.com/publ.html}

UCDP website: \url{https://ucdp.uu.se/downloads/}
}
\usage{
data(nodeatts)
}
\description{
A data frame of six attributes associated with the 733 actors in the global conflict network for the period of 1989-1995. For details on how temporally dynamic variables are aggregated, see Chen (2019). The row ids are Correlates of War country codes or Uppsala Confict Data Program actor codes.
}
\references{
Allansson, M, E Melander, and L Themner, 2017, "Organized violence, 1989-2016." \emph{Journal of Peace Research} 54(4): 727-742. \url{https://doi.org/10.1177/0022343317718773}

Chen, T, 2019, "Statistical Inference for Multilayer Networks in Political Science."

Gleditsch, KS, I Salehyan, and K Schultz, 2008, "Fighting at home, fighting abroad: How civil wars lead to international disputes." \emph{Journal of Conflict Resolution} 52(4): 479-506. \url{https://doi.org/10.1177/0022002707313305}
}
\keyword{datasets}
