% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalconflict-data.R
\docType{data}
\name{nsconf}
\alias{nsconf}
\title{Nonstate Conflict Network}
\format{
A matrix with 544 columns and 544 rows.
}
\source{
UCDP website: \url{https://ucdp.uu.se/downloads/}
}
\usage{
data(nsconf)
}
\description{
An undirected network of conflict between two nonstate actors. This network data is adapted from the Uppsala Conflict Data Program (UCDP) Organized Conflict data set (2017, v17.1). The network includes 544 nonstate actors that were involved in any kind of organized conflict in the 1984-2000 period. The row and column names are actor codes from UCDP.
}
\references{
Allansson, M, E Melander, and L Themner. 2017, "Organized violence, 1989-2016." \emph{Journal of Peace Research} 54(4): 727-742. \url{https://doi.org/10.1177/0022343317718773}
}
\keyword{datasets}
