% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalconflict-data.R
\docType{data}
\name{polrel}
\alias{polrel}
\title{Politically Relevant Network}
\format{A matrix with 189 columns and 189 rows.}
\source{
Gleditsch's website: \url{http://ksgleditsch.com/publ.html}
}
\usage{
data(polrel)
}
\description{
An undirected network where ties indicate the dyad is a politically relevant one. This means that the two countries are physically contiguous or at least one country is a major power. This network data is adapted from Gleditsch, Salehyan, and Schultz (2008). The network includes 189 states in the international system during the post-Cold War period (1989-1995) except Samoa. The row and column names are country codes from the Correlates of War project.
}
\references{
Replication materials for Gleditsch, KS, I Salehyan, and K Schultz, 2008, "Fighting at home, fighting abroad: How civil wars lead to international disputes." \emph{Journal of Conflict Resolution} 52(4): 479-506. \url{https://doi.org/10.1177/0022002707313305}
}
\keyword{datasets}
