% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer_utilities.R
\name{to.multiplex}
\alias{to.multiplex}
\title{To Multiplex}
\usage{
to.multiplex(..., output = c("matrix", "network"), directed = TRUE,
  offzeros = FALSE)
}
\arguments{
\item{...}{matrices or \code{network} objects, or a list of matrices or \code{network} objects.}

\item{output}{one of \code{"matrix"} or \code{"network"}; can be abbreviated.}

\item{directed}{logical; if output is a \code{network} object, whether it should be a directed network.}

\item{offzeros}{logical; if the off diagonal blocks should be all zeros.}
}
\value{
A multiplex network in either matrix or \code{network} object format, depending on whether the output is specified as \code{"matrix"} or \code{"network"}.

If a \code{network} object is returned, it will have the required vertex attribute \code{'layer.mem'} where the layer id is the numeric order in which they were supplied to the function.
}
\description{
Convert supplied network or matrix objects into a multiplex network.
}
\details{
Multiplex networks are a subset of multilayer networks where multiple types of ties exist among the same set of actors, 
represented by different nodes on different layers of the multilayer network. 
The matrices or \code{network} objects supplied to the function must therefore contain the same number of nodes.

Setting \code{offzeros = TRUE} is a convenient way to create matrices used in specifying constraints for multiplex networks.
}
\examples{
mat1 <- matrix(1, nrow = 3, ncol = 3) 
mat2 <- matrix(2, nrow = 3, ncol = 3)
ml.net <- to.multiplex(mat1, mat2, output = "network")
summary(ml.net)

}
\references{
Chen, T, 2019, "Statistical Inference for Multilayer Networks in Political Science."
 
Kivela M et al., 2014, "Multilayer Networks." \emph{Journal of Complex Networks} 2(3): 203-271. \url{https://doi.org/10.1093/comnet/cnu016}
}
