% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemg-data.R
\docType{data}
\name{types}
\alias{types}
\title{ChemG Actor Types}
\format{A character vector of length 30.}
\source{
Dataverse: \url{http://hdl.handle.net/1902.1/17004}
}
\usage{
data(types)
}
\description{
One of five organization types for each of the 30 ChemG actors. The types of organizations are government agencies, interest groups, international organizations, political parties, and scientific organizations.
}
\references{
Replication materials for Leifeld, P and V Schneider, 2012, Information Exchange in Policy Networks. \emph{American Journal of Political Science} 53(3): 731-744. \url{http://dx.doi.org/10.1111/j.1540-5907.2011.00580.x}
}
\keyword{datasets}
