% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer_utilities.R
\name{uptolo}
\alias{uptolo}
\title{Fill Lower Triangular Part of Matrix with its Upper Triangular Part}
\usage{
uptolo(mat, fill.na = TRUE)
}
\arguments{
\item{mat}{a square matrix.}

\item{fill.na}{logical; whether \code{NAs} in the symmetric matrix should be changed to 0s.}
}
\value{
A symmetric matrix.
}
\description{
Utility function that overwrites the lower triangular part of the supplied matrix with the upper triangular part.
}
\examples{
mat <- matrix(c(0,1,1,1,
                0,0,1,1,
                0,0,0,1,
                0,0,0,0),ncol = 4, byrow = T)
mat <- uptolo(mat)
mat
}
